Decouple `ChatMessageListVC` + Introduce`ChatChannelVC`
