+++ DemoApp/CreateChatViewController.swift
@@ -29,7 +29,6 @@ class CreateChatViewController: UIViewController {
                 // Present the new chat and controller
                 let vc = ChatMessageListVC()
                 vc.channelController = controller
-                vc.userSuggestionSearchController = controller.client.userSearchController()
                 
                 navController.setViewControllers([navController.viewControllers.first!, vc], animated: true)
             }
